ref: link contracts function & add post, pre pack, post install
